backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
pydantic>=2.10.0
chromadb>=0.5.0
google-genai>=1.0.0
PyPDF2>=3.0.1
sentence-transformers>=3.3.0
numpy>=1.26.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
firebase-admin>=6.6.0
aiofiles>=24.1.0
httpx>=0.28.0